Support for h11, h2 and qh3 (and remove dependency on http.client) (#1)
Added support for HTTP/1.1, HTTP/2 and HTTP/3 independently of httplib.

Currently, urllib3 does not offer async HTTP requests and the backend is
the http.client package
shipped alongside Python. This implementation is not scheduled to
improve, even less to support the latest
protocols.

Without proxies, the negotiation is as follows:

- HTTP requests are always made using HTTP/1.1.
- HTTPS requests are made with HTTP/2 if TLS-ALPN yields its support
otherwise HTTP/1.1.

- HTTPS requests may upgrade to HTTP/3 if the latest response contains a
valid Alt-Svc header.

With proxies:

- The initial proxy request is always issued using HTTP/1.1 regardless
if it's http or https.
- Subsequent requests follow the previous section (Without proxies)
except that HTTP/3 upgrade is disabled.

You may explicitly disable HTTP/2 or, and, HTTP/3 by passing
``disabled_svn={HttpVersion.h2}`` to your ``BaseHttpConnection``
instance.
Disabling HTTP/1.1 is forbidden and raises an error.

Note that a valid or accepted Alt-Svc header in urllib3 means looking
for the "h3" (final specification) protocol and disallow switching
hostname for security
reasons.