:bookmark: Release 2.5.900 (#78)
2.5.900 (2024-01-31)
====================

- Improved performance and reliability for concurrent streams handled by
a single connection.
We relied on a flat array of events generated by the protocol
state-machine that unfortunately was not
efficient. urllib3-future now handle the events with a
matrix/multi-dimensional array bound to time.
- Fixed a thread safety issue when a single multiplexed connection was
used across many threads.
We revised in-depth the logic wrapper around the connection locking to
make sure you may
go all-in when using threads in that particular context. In consequence
to that:
1) We are, effective immediately, deprecating ``RecentlyUsedContainer``
in favor of our internal ``PoliceTraffic`` that were used in
``PoolManager``.
2) No longer using ``Queue`` to manage the ``Connection`` in
``HTTPConnectionPool``.
If you try to set ``HTTPConnectionPool.QueueCls`` it will raise a
deprecation warning.
Starting today, we no longer accept implementation like ``queue.Queue``
because it
cannot fit the need of this complex HTTP client, especially with the
multiplexing aspect.
- Increased default pool maxsize for DNS-over-HTTPS from 1 to 10.

Close https://github.com/jawah/niquests/issues/65