:bookmark: Release 2.8.904 (#136)
- Relaxed h11 constraint around "pending proposal" and coming server
event about upgrade. This is made to ensure near perfect compatibility
against the legacy urllib3 which is based on http.client.
- Fixed h11 yielding bytearray instead of bytes in rare circumstances.
- Added ``docker-py`` in our CI/integration pipeline.

<!---
Hello!

If this is your first PR to urllib3.future please review the
Contributing Guide:
https://urllib3future.readthedocs.io/en/latest/contributing.html

Adhering to the Contributing Guide means we can review, merge, and
release your change faster! :)
--->