Fix unhandled header as bytes for the unicode content-type transparency (#142)
This PR aims to fix a bug I encountered when using a library that
depends on urllib3.future in the same venv as boto3. This is I believe
the same issue raised in #133.

The reason this occurred is that since urllib3 v2, headers were typed to be string only (k,v). but current usage shows that bytes are still used as values.
Fortunately, the rest of the code already take that into account. The issue was specifically located in the "content-length" x "charset" transparency addition in urllib3-future.
moreover, it seems that no test available in botocore or boto3 cover that specific case, another reason why we missed it.

---------

Co-authored-by: Ahmed TAHRI <ahmed.tahri@cloudnursery.dev>